Revert "Adjust SetupUser to Fchown all open file descriptors so they're appropriately usable by the forked process"
This reverts commit 75150acc2b2baf01a8ce5b106f226464f3c737e0.

There's no guarantee that gosu will be able to change the owner of open
files. Consequently, in some environments, this causes gosu to fail with
the error:

    error: failed switching to "$user": fchown fds permission denied

For some examples of failures people have encountered, see
https://github.com/docker-library/mongo/issues/56

Any open file descriptors may not even be files, or they may not be on a
filesystem that supports changing owners, or even has any concept of
owners (FAT16, for example).

File descriptors may not even be files. They may be sockets, for
example.

This behavior is also quite surprising. I would not expect a utility
that behaves something like sudo to modify my filesystem. Many times, I
may explicitly not want a file to be owned by the target user because I
want the process I'm launching to be restricted to read-only access.
Consequently, this behavior could introduce security issues for some use
cases.